+++ src/components/ImportJobs/ImportJobs.js
@@ -2,8 +2,10 @@ import React, { Component } from 'react';
 import { withRouter, Redirect } from 'react-router';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
+import { get } from 'lodash';
 
 import FileUploader from './components/FileUploader';
+import InvalidFilesModal from './components/InvalidFilesModal';
 
 import css from './components/FileUploader/FileUploader.css';
 
@@ -16,35 +18,62 @@ class ImportJobs extends Component {
 
   state = {
     isDropZoneActive: false,
+    isModalOpen: false,
     redirect: false,
   };
 
   onDragEnter = () => {
-    this.setState({
-      isDropZoneActive: true,
-    });
-  };
+    this.setState({ isDropZoneActive: true });
+  }
 
   onDragLeave = () => {
-    this.setState({
-      isDropZoneActive: false,
-    });
-  };
+    this.setState({ isDropZoneActive: false });
+  }
 
+  /**
+   * @param  {Array<File>} acceptedFiles
+   * @param  {Array<File>} rejectedFiles
+   */
   onDrop = (acceptedFiles, rejectedFiles) => {
-    this.setState({
-      isDropZoneActive: false,
-      acceptedFiles,
-      rejectedFiles,
-      redirect: true,
-    });
+    const isValidFileExtensions = this.validateFileExtensions(acceptedFiles);
+
+    if (isValidFileExtensions) {
+      this.setState({
+        isDropZoneActive: false,
+        redirect: true,
+        acceptedFiles,
+        rejectedFiles,
+      });
+
+      return;
+    }
+
+    this.setState({ isDropZoneActive: false });
+    this.showModal();
   };
 
-  getMessageById = (idEnding, moduleName = 'ui-data-import') => {
+  /**
+   * @param  {Array<File>} files
+   */
+  validateFileExtensions(files = []) {
+    const filesType = get(files, [0, 'type'], '');
+
+    return files.every(({ type }) => type === filesType);
+  }
+
+  showModal() {
+    this.setState({ isModalOpen: true });
+  }
+
+  hideModal = () => {
+    this.setState({ isModalOpen: false });
+  }
+
+  getMessageById(idEnding, moduleName = 'ui-data-import') {
     const id = `${moduleName}.${idEnding}`;
 
     return <FormattedMessage id={id} />;
-  };
+  }
 
   render() {
     const {
@@ -52,6 +81,7 @@ class ImportJobs extends Component {
       rejectedFiles,
       redirect,
       isDropZoneActive,
+      isModalOpen,
     } = this.state;
     const { match } = this.props;
     const titleMessageIdEnding = isDropZoneActive ? 'activeUploadTitle' : 'uploadTitle';
@@ -82,7 +112,15 @@ class ImportJobs extends Component {
         onDragEnter={this.onDragEnter}
         onDragLeave={this.onDragLeave}
         onDrop={this.onDrop}
-      />
+      >
+        {openFileUploadDialogWindow => (
+          <InvalidFilesModal
+            isModalOpen={isModalOpen}
+            onConfirmModal={this.hideModal}
+            openFileUploadDialogWindow={openFileUploadDialogWindow}
+          />
+        )}
+      </FileUploader>
     );
   }
 }