+++ src/components/Report/components/RecordPreview/RecordPreview.js
@@ -1,6 +1,9 @@
 import React from 'react';
 
-import { Row, Col } from '@folio/stripes/components';
+import {
+  Row,
+  Col,
+} from '@folio/stripes/components';
 
 import css from './RecordPreview.css';